created a separate ViewerTooltip component
