+++ app/assets/javascripts/3dbio_viewer/src/webapp/components/BasicInfoViewer.tsx
@@ -4,7 +4,7 @@ import { recordOfStyles } from "../../utils/ts-utils";
 import i18n from "../utils/i18n";
 import { Selection } from "../view-models/Selection";
 import { Links } from "./Link";
-import { Tooltip, ClickAwayListener, Fade } from "@material-ui/core";
+import { ViewerTooltip } from "./viewer-tooltip/ViewerTooltip";
 
 export interface BasicInfoProps {
     pdb: Pdb;
@@ -46,29 +46,20 @@ interface ChildProps {
 const Child: React.FC<ChildProps> = props => {
     const { name, value, help } = props;
     const [showTooltip, setShowTooltip] = useState(false);
-    const handleClose = () => setShowTooltip(false);
-    const handleOpen = () => setShowTooltip(true);
 
     return (
         <li>
             {name}: {value ?? "-"}
             {help && (
-                <ClickAwayListener onClickAway={handleClose}>
-                    <Tooltip
-                        title={help}
-                        placement="right-end"
-                        interactive
-                        TransitionComponent={Fade}
-                        TransitionProps={{ timeout: 600 }}
-                        open={showTooltip}
-                        onOpen={handleOpen}
-                        onClose={handleClose}
-                    >
-                        <span style={styles.help} onClick={() => setShowTooltip(!showTooltip)}>
-                            [?]
-                        </span>
-                    </Tooltip>
-                </ClickAwayListener>
+                <ViewerTooltip
+                    title={help}
+                    showTooltip={showTooltip}
+                    setShowTooltip={setShowTooltip}
+                >
+                    <span style={styles.help} onClick={() => setShowTooltip(!showTooltip)}>
+                        [?]
+                    </span>
+                </ViewerTooltip>
             )}
         </li>
     );
