+++ app/assets/javascripts/3dbio_viewer/src/webapp/components/ViewerBlock.tsx
@@ -1,7 +1,7 @@
 import React, { useState } from "react";
 import _ from "lodash";
 import styles from "./viewers/Viewers.module.css";
-import { Tooltip, ClickAwayListener, Fade } from "@material-ui/core";
+import { ViewerTooltip } from "./viewer-tooltip/ViewerTooltip";
 
 export interface BlockProps {
     block: ViewerBlockModel;
@@ -22,28 +22,18 @@ export const ViewerBlock: React.FC<BlockProps> = React.memo(props => {
     const stringNamespace = _.mapValues(namespace, value => (value || "?").toString());
     const interpolatedDescription = _.template(description)(stringNamespace);
 
-    const handleClose = () => setShowTooltip(false);
-    const handleOpen = () => setShowTooltip(true);
-
     return (
         <div className={styles.section} id={block.id}>
             <div className={styles.title}>
                 {title}
                 {help && (
-                    <ClickAwayListener onClickAway={handleClose}>
-                        <Tooltip
-                            title={help}
-                            placement="right-end"
-                            interactive
-                            TransitionComponent={Fade}
-                            TransitionProps={{ timeout: 600 }}
-                            open={showTooltip}
-                            onOpen={handleOpen}
-                            onClose={handleClose}
-                        >
-                            <button onClick={() => setShowTooltip(!showTooltip)}>?</button>
-                        </Tooltip>
-                    </ClickAwayListener>
+                    <ViewerTooltip
+                        title={help}
+                        showTooltip={showTooltip}
+                        setShowTooltip={setShowTooltip}
+                    >
+                        <button onClick={() => setShowTooltip(!showTooltip)}>?</button>
+                    </ViewerTooltip>
                 )}
             </div>
