+++ app/assets/javascripts/3dbio_viewer/src/webapp/components/viewer-tooltip/ViewerTooltip.tsx
@@ -1,0 +1,35 @@
+import React from "react";
+import { Tooltip, ClickAwayListener, Fade } from "@material-ui/core";
+
+interface ViewerTooltipProps {
+    title: string;
+    showTooltip: boolean;
+    setShowTooltip: (value: React.SetStateAction<boolean>) => void;
+    children: React.ReactElement;
+}
+
+export const ViewerTooltip: React.FC<ViewerTooltipProps> = ({
+    title,
+    showTooltip,
+    setShowTooltip,
+    children,
+}: ViewerTooltipProps) => {
+    const handleClose = () => setShowTooltip(false);
+    const handleOpen = () => setShowTooltip(true);
+    return (
+        <ClickAwayListener onClickAway={handleClose}>
+            <Tooltip
+                title={title}
+                placement="right-end"
+                interactive
+                TransitionComponent={Fade}
+                TransitionProps={{ timeout: 600 }}
+                open={showTooltip}
+                onOpen={handleOpen}
+                onClose={handleClose}
+            >
+                {children}
+            </Tooltip>
+        </ClickAwayListener>
+    );
+};
